requirements.txt
pillow==10.4.0
Django
whitenoise==6.6.0
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-parler===2.3